+++ src/askContainer/webViews/createSkillWebview.ts
@@ -25,6 +25,7 @@ export class CreateSkillWebview extends AbstractWebView {
 
     constructor(viewTitle: string, viewId: string, context: ExtensionContext) {
         super(viewTitle, viewId, context);
+        this.isGlobal = true;
         this.loader = new ViewLoader(this.extensionContext, 'createSkill', this);
     }