+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.1.0 - September 22, 2020
+
+- Adds local simulator features in the toolkit.
+- Adds "What's new" screen in the toolkit, shown to users when extension is
+ updated.
+- Adds a "Contact Alexa" status bar option, to provide easier feedback options
+ to toolkit users.
+- Updated resolution of readme gifs, for better visibility.
+- Fixed the workspace helper to add skills in current workspace, without
+ deleting earlier folders. Fixes #21.
+- Fixes the iconpath on toolkit webviews.
+
 ## 2.0.2 - Sep 03, 2020
 
 - Revert extension main class on package.json file.